improve understanding by playing one clip at a time
This application is a tool to improve the comprehension of a foreign language.

It lets you watch a video or listen to an audio in small pieces of the whished size. Each piece can be seen or listened several times in a row, then while listening you can read the subtitles in the same language as the spoken language and then with translated subtitles.

You can move back and forth a bit at a time or you can play the video or audio from start to finish without interruption, and at any time you can choose to read or hide the original and or the translated subtitles.